+++ demos/html/ckeditor4/webpack.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const CopyPlugin = require("copy-webpack-plugin");
+const webpack = require("webpack");
 
 module.exports = (config, context) => {
   return {
@@ -33,6 +34,12 @@ module.exports = (config, context) => {
             to: path.resolve(__dirname, "dist/ckeditor4"),
           },
         ],
+      }),
+      // Add the DefinePlugin to define process.env variable
+      new webpack.DefinePlugin({
+        'process.env': {
+          CKEDITOR4_API_KEY: JSON.stringify(process.env.CKEDITOR4_API_KEY)
+        },
       }),
     ],
     mode: 'none',