+++ packages/devkit/src/core.src.js
@@ -419,6 +419,8 @@ export default class Core {
    * @param {Node} node - The Node element.
    */
   placeCaretAfterNode(node) {
+    if (node === null) return;
+
     this.integrationModel.getSelection();
     const nodeDocument = node.ownerDocument;
     if (typeof nodeDocument.getSelection !== 'undefined' && !!node.parentElement) {
@@ -570,7 +572,7 @@ export default class Core {
    * @param {HTMLElement} target - The target HTMLElement where formulas should be inserted.
    * @param {Boolean} isIframe - True if the target HTMLElement is an iframe. False otherwise.
    */
-  openModalDialog(target, isIframe, editorOptions = {}) {
+  openModalDialog(target, isIframe) {
 
     // Count the time since the editor is open
     this.editionProperties.editionStartTime = Date.now();
@@ -732,7 +734,6 @@ export default class Core {
     contentManagerAttributes.language = this.language;
     contentManagerAttributes.customEditors = this.customEditors;
     contentManagerAttributes.environment = this.environment;
-    contentManagerAttributes.options = editorOptions || {};
 
     if (this.modalDialog == null) {
       this.modalDialog = new ModalDialog(editorAttributes);